use size_t instead of int

+++ src/CppDistances.h
@@ -57,14 +57,14 @@ std::vector<size_t> CppKNNIndice(
     const std::vector<std::vector<double>>& embedding_space,
     size_t target_idx,
     size_t k,
-    const std::vector<int>& lib,
+    const std::vector<size_t>& lib,
     bool include_self = false);
 
 std::vector<size_t> CppDistKNNIndice(
     const std::vector<std::vector<double>>& dist_mat,
     size_t target_idx,
     size_t k,
-    const std::vector<int>& lib,
+    const std::vector<size_t>& lib,
     bool include_self = false);
 
 std::vector<std::vector<size_t>> CppDistSortedIndice(
